pyproject.toml
[tool.poetry]
name = "inno_assesment"
version = "0.1.0"
description = ""
authors = ["l4blee <l4blee@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^2.0.35"
openai = "^1.46.1"
aiogram = "^3.13.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
PyPDF2 = "^3.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
